backend/requirements.txt
amqp==5.1.1
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
cryptography==41.0.4
daphne==4.0.0
Django==4.2.4
django-cors-headers==4.2.0
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.4
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kombu==5.3.2
msgpack==1.0.7
prompt-toolkit==3.0.39
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.0
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
vine==5.0.0
wcwidth==0.2.8
zope.interface==6.1
